+++ lib/src/core/signal_client.dart
@@ -36,7 +36,6 @@ import '../support/disposable.dart';
 import '../support/platform.dart';
 import '../support/websocket.dart';
 import '../types/other.dart';
-import '../types/video_dimensions.dart';
 import '../utils.dart';
 
 class SignalClient extends Disposable with EventsEmittable<SignalEvent> {
@@ -436,55 +435,10 @@ extension SignalClientRequests on SignalClient {
       ));
 
   @internal
-  void sendAddTrack({
-    required String cid,
-    required String name,
-    required lk_models.TrackType type,
-    required lk_models.TrackSource source,
-    required lk_models.Encryption_Type encryptionType,
-    VideoDimensions? dimensions,
-    bool? dtx,
-    Iterable<lk_models.VideoLayer>? videoLayers,
-    Iterable<lk_rtc.SimulcastCodec>? simulcastCodecs,
-    String? sid,
-    String? stream,
-    bool? disableRed,
-  }) {
-    final req = lk_rtc.AddTrackRequest(
-      cid: cid,
-      name: name,
-      type: type,
-      source: source,
-      encryption: encryptionType,
-      simulcastCodecs: simulcastCodecs,
-      sid: sid,
-      muted: false,
-      stream: stream,
-      disableRed: disableRed,
-    );
-
-    if (type == lk_models.TrackType.VIDEO) {
-      // video specific
-      if (dimensions != null) {
-        req.width = dimensions.width;
-        req.height = dimensions.height;
-      }
-      if (videoLayers != null && videoLayers.isNotEmpty) {
-        req.layers
-          ..clear()
-          ..addAll(videoLayers);
-      }
-    }
-
-    if (type == lk_models.TrackType.AUDIO && dtx != null) {
-      // audio specific
-      req.disableDtx = !dtx;
-    }
-
-    _sendRequest(lk_rtc.SignalRequest(
-      addTrack: req,
-    ));
-  }
+  void sendAddTrack(lk_rtc.AddTrackRequest req) =>
+      _sendRequest(lk_rtc.SignalRequest(
+        addTrack: req,
+      ));
 
   @internal
   void sendUpdateLocalMetadata(lk_rtc.UpdateParticipantMetadata metadata) =>